angular_django_rest_website/BackEnd/requirements.txt
django
djangorestframework
django-cors-headers
Pillow
django-coreapi